Fix crash: integer PCA power iteration (was float/sqrtf, v1.2.1)
The PCA counter crashed the app on push-ups (the only PCA-mode exercise in
that workout; untuned dips and timed holds were fine). Cause: floating-point
math (sqrtf + float projection) in the on-watch hot path — soft-float faulted.

Rewrote the principal-eigenvector solve and the per-sample projection in pure
integer fixed-point: power iteration with shift-renormalisation and isqrt32,
producing a Q12 unit direction; projection is int64 with a >>12. No math.h,
no float anywhere in rep_counter.c. Validated against the Python reference
(2-rep disagreement / 1651) — the integer power iteration converges to the
same direction as numpy's eigh.

+++ src/c/rep_counter.h
@@ -47,7 +47,7 @@ typedef struct {
   // (x,y,z) over the selection window, then its principal eigenvector = the rep
   // direction. Projecting onto it is invariant to how the watch is oriented.
   int64_t cov[6];                // xx,yy,zz,xy,xz,yz (band-passed, in mG^2)
-  float pca_u[3];                // locked principal direction (unit vector)
+  int32_t pca_u[3];              // locked principal direction, Q12 unit vector
   int32_t amp_est_q8;            // running swing-amplitude estimate
   bool in_low;                   // inside a down-swing, waiting for the rise
   int32_t trough_q8;             // deepest point of the current down-swing
